[Bug #7915] Newsletter page using MC template
Created the controller that allows for the liberation of the desired
newsletter after clicking on the "Accept this newsletter" in the summary
email.
Added a AddNewsletterToWhitelist action to the SOAP (Soaper) interface.
Added a newsletter template for the landing page
Got the texts ready for translation through weblate
Modified `send_summary.pl` to get the link to the new page