[#8640] Fixing the User/Add API function
In the case when a local user is added through the API on a domain
configured in LDAP, an error was raised and the action did not complete.
This was due to the behaviour of the function `loadOldDomain` in
`www/guis/admin/application/models/Domain.php` which returned the old
Domain version if it did not exist, but returned nothing when it did.
The fix is simply to return the old Domain in both cases.