Fix db synchronization issues at start
As the MailCleaner is now, the Binary logs for the mysql replication
expire after 21 days. Thus when starting a new VM more than 21 days
after the VM creation, there will be a synchronization issue between the
databases.
Calling the `resync_db.sh` script on the first run allows to solve this
issue.