Fix the db sync on first start
When starting the VA for the first time, if it is more than 21 days
after the VA creation, the binary logs are rotated, making it so that
the slave cannot find the file.
To solve this, the first db start is done by keeping all the binary
logs, then keeping only the last 21 days.