Fix of bug from ticket 10003
When using merged quarantine display, if the summary recipient address in
`Configuration > Addess Settings` was set to other, it will appear as
the current address being modified instead of the custom one.

The part of code removed was using the user's `summary_to` field which
is never set in user configurations, only in emails configurations, and
causing this issue.