Fix not being able to release attachement
When a mail is sent from an Exchange server, Exim registers the
Message-ID on two lines. When trying to resend the email to an other
Exchange server, this last on sees the same Message-ID, since it could
not be changed as in other messages, the ID being on two lines, and
drops the email as a DUPLICATEDELIVER.

When parsing the Headers file, the `force_quarantine.pl` script will
now look for multiple lines Message-IDs, set it on a single line and
adds a suffix to make sure the mail will be accepted by the distant
server.