Add shorthands to hostlist elements
Provides expansion of DNS records to lists of IP blocks for A, AAAA, MX, and SPF (pass), using a suffix with /type

eg. adding _spf.google.com/spf to a trusted ip list will expand to all IPs in the SPF record when the config file is dumped

Also provides a script (bin/dns_lookup.pl) for testing, easy SPF validation, etc.

Also standardizes the dumpers, fixes a couple of bugs related to '/mx' being a pre-existing shorthand with no functionality and IPv6 support in SNMP.