Alternate whitelisting options
Matches variations of an address, including sub-addresses, SRS, etc. for wwlist rules

Offers the ability to create a rule for the simplified address if it is one of the supported variations

Offeres the ability to create a domain-wide rule if the simplified address still looks like a single-use address.

Fixes a the newslist button in the user quarantine to also offer these variations.