Remove newsletter flag when allowed, better logging
If the newsletter is newslisted or the user allows newsletters then remove this flag before storing it to the DB. This will prevent messages from indicating that they were detected as Newsletters in the WebUI to avoid confusion and provide the more intuitive Whitelisting pop-up when released.

Improved the log output and SpamCheck header for whitelisted and/or newslisted items to be more cleare about which was seen at which level.