Fix failure to bypass for recipients with no filtering
As of Exim 4.94 we need to use the sanitized $domain_data instead of $domain. However, unlike $domain, $domain_data will retain resulting values from a lookup. This means that for a lookup like:

  wildlsearch;VARDIR/spool/tmp/mailcleaner/domains.list

which contains:

  domain.com:  destination.server:25 byname randomize

$domain_data would contain the value instead of the key (domain name), resulting in the wrong file being looked up for later searching using $domain_data.

To resolve this we need to specify that the lookup should return the key instead:

  wildlsearch;ret=key;VARDIR/spool/tmp/mailcleaner/domains.list