Prevent double quarantine for newsl and bl
Stage 4 has a redundant check for blacklisted senders which could have bypassed the filtering stage. To ignore a blacklist result, it must already have a Blacklisted header.

The problem is that messages that are quarantined for other reasons, or a sender that was blacklisted after the message was quarantined will not have this header either. This results in Stage 4 adding the header to a released message and re-quarantining it. This fix simply checks if the message was forced and will also ignore blacklists during Stage 4 for that reason.