Prevent errors for unresolvable hostnames
PreRBLs and UriRBLs (and others?) accept '*.domain.tld' format, but this broke the IP math functions.

Push any type of hostname to a separate list and concatenate them with the main list after expansion and simplification is over.

Also disables debug mode. This is hidden by the init scripts anyways, but best to avoid confusion if the dumpers are run manually.