Move full whitelisted hostlist to dedicated file
The file is full_whitelisted_hosts.list, in the same directory as full_whitelisted_senderss.list

If this file does not exist, continue to use the trusted_ips list to support legacy setups.

This file is dumped with the dns expansion module, so it supports the exact same formats as trusted_ips.

This is preferable to continued use of trusted_ips because the latter requires that all other (not fully whitelisted) senders from those IPs must skip spam checking even if you don't want them to.