bug fixes for user wwlist forms (#383)
It was possible to add duplicate entries which could not be deleted.

This changes to check the status of the checkbox variable since the name variable holding an array always gets set to 0.

Also checks existing entries before allowing for duplicates to be added in the first place.

A limitation of the form processing means that it is not trivial to delete more than one of the duplicates at a time, so they need to be deleted individually, if necessary. Since duplicates should not be allowed, this is only an issue for legacy rules.