Add TrustedIPs prefilters module (#409)
Accompanies matching U4MC commit

Takes the evaluation of TrustedIPs and White-IP-Dom (domain IP whitelist) out of the TrustedSources module and creates a unique module for them.

This allows these to be enforced even if TrustedSources is disabled.

This module is hidden in the AntiSpam menu since the settings are located elsewhere already. It has the highest priority to run before all other modules.

The configuration file is static since it does not contain any variables and these whitelists cannot be disabled.