Add an 'Unsubscribe' using List-Unsubscrbe header
Simplest implementation first. The message preview already loads the message headers. Simply load the cleaned header content into the object and if it is assigned then show a link next to the existing newslist link.

Other locations are more complicated because the headers are not searched and the unsubscribe link is not logged in the db (yet?)