IP RBL checking tool
run without arguments for usage.

In short: it will extract the Received IPs from a .eml or take a raw IPv4 address, and check them against any RBL enabled in SMTP, PreRBLs or SpamC, then output any hits, along with the levels at which that source is enabled.

Useful for diagnosing a quarantined message where either it hit a IN_RCVD rule or to see if another listing still exists.