Fix per-domain SpamC score overrides
SpamAssassin will "merge" identical rules, which means that only the first will be retained. As a result, for component rules generated for the same list of domains, only the first will be valid.

  meta __RULE1_RCPT_LIST ( __RCPT1 || __RCPT2 ) <- used
  meta __RULE2_RCPT_LIST ( __RCPT1 || __RCPT2 ) <- discarded

It does not create an alias, so any meta rule referencing the latter cannot hit.

Having the additional component rule for the recipient list was not necessary in the first place, since they were duplicated for every rule anyways. So, the fix is to simply not create any of these 'recipient list' component rules and instead just include the same '||' clause within the final meta rule.