UNTESTED: folding long lines in exim
4.95 started to strictly enforce line length limits (998 characters).
This is good except that Outlook (and others) fail to adhere to RFC.

This adds a folding.pl script to a transport_filter which will transform
any lines over that length by first, breaking at spaces, then breaking at
commas, then breaking mid-word.

Adds a DB column for mta_config to make this opt-in. This is because it
could break DKIM signing.

Totally untested outside of feeding a couple of .eml files directly to the
script, so DO NOT MERGE YET.