Working implementation of line limit fix
'allow_long' configuration was added. This is enabled by default and will allow the limit to be ignored.

Nested inside this condition on each transport is the previous 'folding' option (disabled by default).

The default action will therefore be to ignore the bad SMTP and pass it on (still as bad SMTP).

Users can then opt-in to folding which will make the SMTP valid if their back-end complains. Outgoing DKIM signing will be done after the folding, so it will be valid. Relayed inbound messages can have DKIM broken by other headers anyways, but this makes it more likely.

If both are disabled, there is a new ACL which will reject the message with a proper 5xx code instead of just presenting the MTA error.