Use custom init manager for Fail2Ban
Potential fix for Fail2Ban restart problem with large blacklist.

On our Cloud cluster Fail2Ban can fail to restart if it is restarted in quick succession. This is because the initial startup is not done starting when it tries to stop, then cannot start again because the first iteration is locking the socket.

The custom init library that I created a couple of years ago properly checks to see if the process is actually still running and will prevent this blocking.

It is possible to resolve this in the existing init script, but this also allows for Fail2Ban to be added to the CheckAll function which means that it can be restarted automatically if all else fails.