Swap 'Version' value with 'Patch level' value.
These values were previously redundant. The 'Version' has been changed be meaningful by including the installed patch number and the Git commit hash. This is done in U4MC.

The patch level, eg. "2023100301", makes more sense as a version number, and vice versa (eg. 83-03dd071). This commit simply swaps the values associated with each label in the WebUI. It does not change the actual usage of these values on the back-end, since the version gets used in the SMTP banner and in that case, the commit hash and installed updates are more meaningful.