Reindex script (#607)
* WIP: Script to re-index logs

Search all existing log files and generate a new list of files with dates properly segmented into proper filenames. This is necessary when log rotation fails or is run too often or at the wrong times.

Requires optimization, especially for "fast" mode, which should be optimized for speed at the cost of memory, by ingesting all logs of a given type into a hash with a key for each file. At the moment, it does way more checks than it needs to for each line, so it is generally slightly slower than the normal mode.

It is unlikely that this will ever be fast enough to warrant running proactively in Cron.

* Replace Kaspersky with ESET in add-on options (#605)

* Add key for Frederic

* Update key removal during unregistration

* Fixes to Fast mode

Requires testing and verification that Slow mode still works.

* Fix permissions and don't bother writing empty log