Mailscanner eximid (#616)
* Add option defining Exim path

This will be used with accompanying patch to MailScanner so that longer Exim IDs in 4.97+ can be properly dealt with.

* Dynamically add Exim Command option if supported

MailScanner presents an error if it is not yet upgraded to expect this option. To keep things clean in the event that a machine is behind on updates, we will instead only add the option if it is supported in ConfigDefs.pl for the local installation.

* Add option defining Exim path

This will be used with accompanying patch to MailScanner so that longer Exim IDs in 4.97+ can be properly dealt with.

* Dynamically add Exim Command option if supported

MailScanner presents an error if it is not yet upgraded to expect this option. To keep things clean in the event that a machine is behind on updates, we will instead only add the option if it is supported in ConfigDefs.pl for the local installation.