New cdn (#622)
* Fix log permissions

* Update to AWS bucket

* Update CDN path

* Completely rework Bayes fetcher

Needed to change for AWS as CDN.

Rather than Enterprise users downloading HTTP credentials, it now downloads a secret (the md5 hash of the target file). If secrets are downloaded, that has is used to locate the target file, otherwise (Community edition) it will download a monthly snapshot from a public path.

Accompanies changes to infrastructure scripts.