Exim 4.96 and onward have compatible templates
Upgrade the generic template to the latest and remove redundant templates since 4.94 is not available in the repositories anymore and we have not included it in VM images for years. Watchdogs report that no Enterprise edition users still have versions prior to 4.96. If there are community edition users, they should have been receiving alerts for years. If they are not receiving alerts, it is because their appliance is already failing to update, so this commit will not be pulled.