Safely fall back from missing/invalid SpamCheck header
If there is a 'spam, ' result in the header but without a known module, it may cause the message to not be rejected as spam. This will log the unknown module but trust the 'spam' result in order to quarantine/tag/whitelist, etc.

As a fallback to that fallback, this will also remove the `{MC_SPAM}` subject tag if it were to somehow exist for a ham. If this is left intact but SpamHandler does not detect it as spam, it will loop with Stage 4 because SendMeAnyway will add it back to the Stage 4 queue, then the `{MC_SPAM}` subject will pass it back to SpamHandler.