Start of 'lazy migration' to Bookworm
This is some work towards a less ambitious migration where several of
the more complex migration goals are being rolled back. This aims to
take any of the enhancements that work, then try to push out everything
else in as close to the form that it takes in the Jessie release as
possible.

One project which is being abandoned is reproducible builds. Because of
this the 'debian_bootstrap' method is being discarded. This repository
is being applied to an existing Jessie image which has been migrated
from Jessie -> Stretch -> Buster -> Bullseye -> Bookworm. However, a lot
of other changes have to be made to get it to the state that I'm testing
at. Especially:

Lots of symlinking, removing of existing packages, installing additional
packages, removing disabling some init services, and much more. Instead
of reproducible builds, we will aim to have a golden image which will
continue to be updated and distributed as snapshots.

Not everything works yet, most especially the WebUI. Pushing at the
current state will allow me to have the commit open in another window to
more easily see what has changed and what might still need to change.

These changes have been pulled in piecemeal from MailCleaner-Next and
modified as necessary. Many files will need to be reverted to go back to
the Jessie way of doing things and other files have yet to be pulled in.